+++ include/NumCpp/Functions/geomspace.hpp
@@ -28,6 +28,12 @@
 #pragma once
 
 #include "NumCpp/NdArray.hpp"
+#include "NumCpp/Core/Internal/Error.hpp"
+#include "NumCpp/Core/Internal/StaticAsserts.hpp"
+#include "NumCpp/Functions/logb.hpp"
+#include "NumCpp/Functions/logspace.hpp"
+#include "NumCpp/Functions/nth_root.hpp"
+#include "NumCpp/Utils/essentiallyEqual.hpp"
 
 namespace nc
 {
@@ -48,5 +54,28 @@ namespace nc
     /// @param enpoint: If true, stop is the last sample. Otherwide,it is not included. Default is true.
     /// @return NdArray
     ///
+    template<typename dtype>
+    NdArray<double> geomspace(dtype start, dtype stop, uint32 num = 50, bool endPoint = true)
+    {
+        STATIC_ASSERT_ARITHMETIC_OR_COMPLEX(dtype);
 
+        if (utils::essentiallyEqual(start, dtype{0}))
+        {
+            THROW_INVALID_ARGUMENT_ERROR("Geometric sequence cannot include zero");
+        }
+
+        if (num == 1)
+        {
+            return { static_cast<double>(start) };
+        }
+        else if (num == 2 && endPoint)
+        {
+            return {static_cast<double>(start), static_cast<double>(stop)};
+        }
+
+        const auto base = nth_root(stop / start, num - 1);
+        const auto logStart = logb(start, base);
+        const auto logStop = logb(stop, base);
+        return logspace(logStart, logStop, num, endPoint, base);
+    }
 }  // namespace nc