+++ include/NumCpp/Functions/isposinf.hpp
@@ -28,6 +28,7 @@
 #pragma once
 
 #include "NumCpp/NdArray.hpp"
+#include "NumCpp/Core/Internal/StaticAsserts.hpp"
 #include "NumCpp/Functions/isinf.hpp"
 
 namespace nc